feat: add ResultRenderer for handling result output in CLI
