+++ packages/cli/src/renderers/index.ts
@@ -1,4 +1,5 @@
 export * from "./experimental-trajectory-stream-renderer";
 export * from "./json-renderer";
 export * from "./output-renderer";
+export * from "./result-renderer";
 export * from "./types";
