+++ packages/cli/src/renderers/result-renderer.ts
@@ -1,0 +1,41 @@
+import { isToolUIPart } from "ai";
+import type { NodeChatState } from "../livekit/chat.node";
+import type { StreamRenderer } from "./types";
+
+export interface ResultRendererOptions {
+  attemptCompletionSchemaOverride?: boolean;
+}
+
+export class ResultRenderer implements StreamRenderer {
+  private attemptCompletionSchemaOverride: boolean;
+
+  constructor(
+    private readonly stream: NodeJS.WritableStream,
+    private readonly state: NodeChatState,
+    options: ResultRendererOptions = {},
+  ) {
+    this.attemptCompletionSchemaOverride =
+      !!options.attemptCompletionSchemaOverride;
+  }
+
+  async shutdown() {
+    const messages = this.state.signal.messages.value;
+    const lastMessage = messages.at(-1);
+
+    if (lastMessage?.role === "assistant") {
+      for (const part of lastMessage.parts || []) {
+        if (isToolUIPart(part) && part.type === "tool-attemptCompletion") {
+          if (part.input) {
+            const input = part.input as Record<string, unknown>;
+            if (this.attemptCompletionSchemaOverride) {
+              this.stream.write(`${JSON.stringify(input.result, null, 2)}\n`);
+            } else {
+              this.stream.write(`${input.result}\n`);
+            }
+          }
+          return;
+        }
+      }
+    }
+  }
+}
